Syndication zip cache.

+++ src/main/java/org/snomed/snowstormlite/service/AppSetupService.java
@@ -60,6 +60,9 @@ public class AppSetupService {
 
 					importService.importRelease(filePaths, contentItemVersion);
 					for (String filePath : filePaths) {
+						if (syndicationClient.retainsRf2ZipAfterImport(filePath)) {
+							continue;
+						}
 						if (!new File(filePath).delete()) {
 							logger.info("Failed to delete temp file {}", filePath);
 						}
